Split MegaMenu content rendering
